[LETS-68] Fix -Wdeprecated-declarations caused by use of ftime (#2646)
‘int ftime(timeb*)’ is deprecated and should be replaced with the usage of functions from std::chrono

+++ src/broker/broker_util.c
@@ -26,6 +26,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
+#include <chrono>
 #include <fcntl.h>
 #include <string.h>
 #include <errno.h>
@@ -422,7 +423,7 @@ ut_time_string (char *buf, struct timeval *time_val)
 {
   struct tm tm, *tm_p;
   time_t sec;
-  int millisec;
+  long millisec;
 
   if (buf == NULL)
     {
